+++ src/components/monaco/MonacoEditorWrapper.tsx
@@ -3,19 +3,7 @@ import React, { PureComponent } from 'react';
 import MonacoEditor, { MonacoEditorProps } from 'react-monaco-editor';
 import { Themeable, withTheme } from '@grafana/ui';
 
-export interface MonacoLanguageDefinition {
-    /** unique language ID.
-     * Note: there can be multiple Monaco Editors on the same page
-     * where each of them needs a different auto-completion
-     * (each query can use a different datasource or overriden URL)
-     * so we need to make every language definition ID unique */
-    languageId: string;
-    register: () => void;
-    deregister: () => void;
-}
-
 export interface MonacoEditorWrapperProps extends Omit<MonacoEditorProps, 'theme'> {
-    languageDefinition: MonacoLanguageDefinition;
     alwaysShowHelpText?: boolean;
     initMonaco?: (monaco: typeof Monaco) => void;
     onBlur?: (value: string) => void;
@@ -39,16 +27,11 @@ class MonacoEditorWrapper extends PureComponent<Props> {
         this.props.onBlur?.(this.editor?.getValue() || '');
     };
 
-    editorWillMount = (monaco: typeof Monaco) => {
-        this.props.languageDefinition.register();
-        this.props.editorWillMount?.(monaco);
-    };
-
     editorDidMount = (editor: Monaco.editor.IStandaloneCodeEditor, monaco: typeof Monaco) => {
         this.editor = editor;
 
         if (this.props.onSave) {
-            editor.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.KEY_S, () => {
+            editor.addCommand(monaco.KeyMod.CtrlCmd | monaco.KeyCode.KeyS, () => {
                 this.props.onSave?.(editor.getValue());
             });
         }
@@ -56,14 +39,9 @@ class MonacoEditorWrapper extends PureComponent<Props> {
         this.props.editorDidMount?.(editor, monaco);
     };
 
-    componentWillUnmount() {
-        this.props.languageDefinition.deregister();
-    }
-
     render() {
         const props: MonacoEditorProps = {
             ...this.props,
-            language: this.props.languageDefinition.languageId,
             theme: this.props.theme.isDark ? 'vs-dark' : 'vs-light',
             options: {
                 wordWrap: 'on',
@@ -99,7 +77,7 @@ class MonacoEditorWrapper extends PureComponent<Props> {
 
         return (
             <div onBlur={this.onBlur}>
-                <MonacoEditor editorWillMount={this.editorWillMount} editorDidMount={this.editorDidMount} {...props} />
+                <MonacoEditor editorDidMount={this.editorDidMount} {...props} />
             </div>
         );
     }