+++ src/components/monaco/index.ts
@@ -1,0 +1,4 @@
+import type * as MonacoType from 'monaco-editor/esm/vs/editor/editor.api';
+
+export type { MonacoType };
+export declare type Monaco = typeof MonacoType;