+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **redis**: fix label_names() function to return all label names if no parameter is specified (now the label name auto-completion in the query editor works again)
 - **redis**: remove deprecated `label_values(metric, label)` function
 - **chore**: upgrade Grafana dependencies to version 8.4.1
+- **chore**: refactor custom Monaco languages
 
 ### Removed features
 - **redis**: The `label_values(metric, label)` Grafana variable query function is now removed (was deprecated since grafana-pcp v3)